Refactoring to different files
