+++ provider/adapters/Cargo.toml
@@ -24,10 +24,11 @@ include = [
 ]
 
 [dependencies]
-icu_provider = { version = "0.6.0", path = "../core" }
+icu_provider = { version = "0.6.0", path = "../core", features = ["macros"] }
 icu_locid = { version = "0.6.0", path = "../../components/locid", features = ["zerovec"] }
 yoke = { version = "0.6.0", path = "../../utils/yoke" }
 zerovec = { path = "../../utils/zerovec", features = ["yoke"] }
+serde = { version = "1.0", default-features = false, features = ["derive", "alloc"], optional = true }
 
 [features]
 std = ["icu_locid/std"]
